+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 <!-- id='changelog-0.0.3'-->
-# 0.0.3 — 2024-01-25
+# 0.0.3 — 2024-01-26
 
 ## Added
 
 - added a mechanic initially adding all connected coordinates to a Set GraphMatrix.matrix_coords
 - added List and Dictionary as return types for functions offering optional boolean choice to return RectCoords or HexCoords, selection is now based on entering the desired return data type as a String, the options being: "List", "Dict", "Tuple", "Coords" (for RectCoords or HexCoords, depending on returned length)
 - added 64_px_example_tilemask.png
+- added assertequals for dict and list output in unittests
+- GraphMatrix based pathfinding returns None if no path from start to goal was found
+- optional parameter for GraphMatrix based pathfinding, test_accessibility, tests if start and goal are connected to other tiles, if not returns None
 
 ## Changed